requirement.txt
fastapi
uvicorn
torch
transformers
openai-whisper
numpy
ffmpeg-python  # Required for Whisper
